Refactor widgets into modules
