+++ ui/sprig_ui/lib/main.dart
@@ -1,286 +1,6 @@
 import 'package:flutter/material.dart';
-import 'package:sprig_ui/repo.dart';
-import 'package:sprig_ui/split.dart';
+import 'package:sprig_ui/screens/home.dart';
 
 void main() {
   runApp(const SprigUI());
-}
-
-class SprigUI extends StatelessWidget {
-  const SprigUI({super.key});
-
-  // This widget is the root of your application.
-  @override
-  Widget build(BuildContext context) {
-    return MaterialApp(
-        title: 'Sprig',
-        theme: ThemeData(
-          // This is the theme of your application.
-          //
-          // TRY THIS: Try running your application with "flutter run". You'll see
-          // the application has a purple toolbar. Then, without quitting the app,
-          // try changing the seedColor in the colorScheme below to Colors.green
-          // and then invoke "hot reload" (save your changes or press the "hot
-          // reload" button in a Flutter-supported IDE, or press "r" if you used
-          // the command line to start the app).
-          //
-          // Notice that the counter didn't reset back to zero; the application
-          // state is not lost during the reload. To reset the state, use hot
-          // restart instead.
-          //
-          // This works for code too, not just values: Most code changes can be
-          // tested with just a hot reload.
-          colorScheme: ColorScheme.fromSeed(seedColor: Colors.green),
-          useMaterial3: true,
-        ),
-        home: Scaffold(
-          appBar: AppBar(
-            // TRY THIS: Try changing the color here to a specific color (to
-            // Colors.amber, perhaps?) and trigger a hot reload to see the AppBar
-            // change color while the other colors stay the same.
-            backgroundColor: Theme.of(context).colorScheme.inversePrimary,
-            // Here we take the value from the MyHomePage object that was created by
-            // the App.build method, and use it to set our appbar title.
-            title: const Text('Sprig'),
-          ),
-          body: const BasketUI(),
-        ));
-  }
-}
-
-class SprigDetailsCard extends StatelessWidget {
-  const SprigDetailsCard({super.key, required this.sprigDetails});
-
-  final SprigDetails sprigDetails;
-
-  // ListTile(
-  //   leading: const Icon(Icons.data_object_outlined),
-  //   title: Text('${sprigDetails.name}'),
-  //   subtitle: Column(
-  //       crossAxisAlignment: CrossAxisAlignment.start,
-  //       children: <Widget>[
-  //         Text('Structure: ${sprigDetails.structure}'),
-  //         Text('Format: ${sprigDetails.format}'),
-  //         Text('ID: ${sprigDetails.id}')
-  //       ]),
-  // ),
-
-//
-
-  @override
-  Widget build(BuildContext context) {
-    return Card(
-        child: Padding(
-            padding: const EdgeInsets.all(16),
-            child: Column(
-                crossAxisAlignment: CrossAxisAlignment.start,
-                children: <Widget>[
-                  Text('${sprigDetails.name}',
-                      style: const TextStyle(fontWeight: FontWeight.bold)),
-                  Text('Structure: ${sprigDetails.structure}'),
-                  Text('Format: ${sprigDetails.format}'),
-                  Text('ID: ${sprigDetails.id}')
-                ])));
-  }
-}
-
-class SprigDetailsPanel extends StatefulWidget {
-  const SprigDetailsPanel({super.key, required this.sprigName});
-
-  final Sprig sprigName;
-
-  @override
-  State<SprigDetailsPanel> createState() => _SprigDetailsPanelState();
-}
-
-class _SprigDetailsPanelState extends State<SprigDetailsPanel> {
-  // FIXME: This default value is just for debugging
-  Basket repo = LocalBasket(
-      sprigBinary: "/Users/dorran/dev/sprig/clients/python/.venv/bin/sprig");
-
-  @override
-  Widget build(BuildContext context) {
-    final asyncSprigWidget = FutureBuilder<SprigDetails>(
-      future: repo.getDetails(widget.sprigName),
-      builder: (BuildContext context, AsyncSnapshot<SprigDetails> snapshot) {
-        List<Widget> children;
-        if (snapshot.hasData &&
-            !(snapshot.connectionState == ConnectionState.waiting)) {
-          children = <Widget>[SprigDetailsCard(sprigDetails: snapshot.data!)];
-        } else if (snapshot.hasError) {
-          children = <Widget>[
-            const Icon(
-              Icons.error_outline,
-              color: Colors.red,
-              size: 60,
-            ),
-            Padding(
-              padding: const EdgeInsets.only(top: 16),
-              child: Text('Error: ${snapshot.error}'),
-            ),
-          ];
-        } else {
-          children = const <Widget>[
-            SizedBox(
-              width: 60,
-              height: 60,
-              child: CircularProgressIndicator(),
-            ),
-            Padding(
-              padding: EdgeInsets.only(top: 16),
-              child: Text('Awaiting result...'),
-            ),
-          ];
-        }
-        final leftPanel = Column(
-          mainAxisAlignment: MainAxisAlignment.center,
-          children: children,
-        );
-
-        return Center(child: leftPanel);
-      },
-    );
-    return asyncSprigWidget;
-  }
-}
-
-/// The main UI component for managing Sprig baskets.
-class SprigList extends StatefulWidget {
-  const SprigList({super.key, required this.onSprigSelected});
-  final Function(Sprig?)? onSprigSelected;
-  @override
-  State<SprigList> createState() => _SprigListState();
-}
-
-class _SprigListState extends State<SprigList> {
-  // FIXME: This default value is just for debugging
-  Basket repo = LocalBasket(
-      sprigBinary: "/Users/dorran/dev/sprig/clients/python/.venv/bin/sprig");
-
-  int? _selectedSprigIndex;
-
-  @override
-  Widget build(BuildContext context) {
-    final asyncSprigWidget = FutureBuilder<Sprigs>(
-      future: repo.list(),
-      builder: (BuildContext context, AsyncSnapshot<Sprigs> snapshot) {
-        List<Widget> children;
-        if (snapshot.hasData) {
-          children = <Widget>[
-            Text(
-              'Sprigs:',
-              style: Theme.of(context).textTheme.titleMedium,
-              textAlign: TextAlign.start,
-            ),
-            Flexible(
-                child: ListView.builder(
-                    padding: const EdgeInsets.all(8),
-                    itemCount: snapshot.data?.sprigs?.length ?? 0,
-                    itemBuilder: (BuildContext context, int index) {
-                      return ListTile(
-                        onTap: () {
-                          // Update the state of this widget
-                          setState(() {
-                            _selectedSprigIndex = index;
-                          });
-                          // Fire off any provided callbacks as well
-                          widget.onSprigSelected
-                              ?.call(snapshot.data?.sprigs?[index]);
-                        },
-                        tileColor: _selectedSprigIndex == index
-                            ? Color.fromARGB(255, 148, 243, 154)
-                            : null,
-                        shape: RoundedRectangleBorder(
-                            side: BorderSide(color: Colors.grey, width: 0.5),
-                            borderRadius: BorderRadius.circular(4)),
-                        leading: const Icon(Icons.data_object_outlined),
-                        title: Text('${snapshot.data?.sprigs?[index].name}'),
-                      );
-                    })),
-          ];
-        } else if (snapshot.hasError) {
-          children = <Widget>[
-            const Icon(
-              Icons.error_outline,
-              color: Colors.red,
-              size: 60,
-            ),
-            Padding(
-              padding: const EdgeInsets.only(top: 16),
-              child: Text('Error: ${snapshot.error}'),
-            ),
-          ];
-        } else {
-          children = const <Widget>[
-            SizedBox(
-              width: 60,
-              height: 60,
-              child: CircularProgressIndicator(),
-            ),
-            Padding(
-              padding: EdgeInsets.only(top: 16),
-              child: Text('Awaiting result...'),
-            ),
-          ];
-        }
-        final leftPanel = Column(
-          mainAxisAlignment: MainAxisAlignment.center,
-          children: children,
-        );
-
-        return Center(child: leftPanel);
-      },
-    );
-    return asyncSprigWidget;
-  }
-}
-
-/// The main UI component for managing Sprig baskets.
-class BasketUI extends StatefulWidget {
-  const BasketUI({super.key});
-
-  @override
-  State<BasketUI> createState() => _BasketUIState();
-}
-
-class _BasketUIState extends State<BasketUI> {
-  /// The currently selected sprig
-  Sprig? selectedSprig;
-
-  @override
-  Widget build(BuildContext context) {
-    callback(selection) {
-      setState(() {
-        selectedSprig = selection;
-      });
-    }
-
-    if (selectedSprig != null) {
-      return Split(
-        axis: Axis.horizontal,
-        initialFractions: const [0.3, 0.7],
-        splitters: [
-          SizedBox(
-            width: 6,
-            child: DecoratedBox(
-              decoration: BoxDecoration(
-                color: Theme.of(context).dividerColor,
-              ),
-            ),
-          ),
-        ],
-        children: [
-          SprigList(onSprigSelected: callback),
-          SprigDetailsPanel(
-            sprigName: selectedSprig!,
-          )
-        ],
-      );
-    } else {
-      return Center(
-          child: SprigList(
-        onSprigSelected: callback,
-      ));
-    }
-  }
 }
