+++ ui/sprig_ui/lib/widgets/sprig_details_card.dart
@@ -1,0 +1,24 @@
+import 'package:flutter/material.dart';
+import 'package:sprig_ui/repo.dart';
+
+class SprigDetailsCard extends StatelessWidget {
+  const SprigDetailsCard({super.key, required this.sprigDetails});
+
+  final SprigDetails sprigDetails;
+
+  @override
+  Widget build(BuildContext context) {
+    return Card(
+        child: Padding(
+            padding: const EdgeInsets.all(16),
+            child: Column(
+                crossAxisAlignment: CrossAxisAlignment.start,
+                children: <Widget>[
+                  Text('${sprigDetails.name}',
+                      style: const TextStyle(fontWeight: FontWeight.bold)),
+                  Text('Structure: ${sprigDetails.structure}'),
+                  Text('Format: ${sprigDetails.format}'),
+                  Text('ID: ${sprigDetails.id}')
+                ])));
+  }
+}
